+++ xmake/core/project/project.lua
@@ -108,6 +108,11 @@ function project._api_has_config(interp, ...)
     return config.has(...)
 end
 
+-- get config from the given name
+function project._api_get_config(interp, name)
+    return config.get(name)
+end
+
 -- add module directories
 function project._api_add_moduledirs(interp, ...)
     sandbox_module.add_directories(...)
@@ -200,7 +205,8 @@ function project.interpreter()
         ,   {"is_plat",                 project._api_is_plat          }
         ,   {"is_arch",                 project._api_is_arch          }
         ,   {"is_config",               project._api_is_config        }
-        ,   {"has_config",              project._api_has_config      }
+        ,   {"get_config",              project._api_get_config       }
+        ,   {"has_config",              project._api_has_config       }
             -- add_xxx
         ,   {"add_moduledirs",          project._api_add_moduledirs   }
         ,   {"add_plugindirs",          project._api_add_plugindirs   }