+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### New features
 
 * Support fasm assembler
-* Add `has_config` and `is_config` apis
+* Add `has_config`, `get_config` and `is_config` apis
 * Add `$ xmake --try` to try building project using third-party buildsystem
 * Add `set_enabled(false)` to disable target 
 
@@ -485,7 +485,7 @@
 ### 新特性
 
 * 新增fasm汇编器支持
-* 添加`has_config`和`is_config`接口去快速判断option和配置值
+* 添加`has_config`, `get_config`和`is_config`接口去快速判断option和配置值
 * 添加`$ xmake --try`去尝试构建工程
 * 添加`set_enabled(false)`去显示的禁用target